+++ 05a_Composite/Cd.cs
@@ -7,7 +7,7 @@ namespace a_Composite
 {
 	public class Cd:Item
 	{
-		public Cd (string name, float price):base(name,price)
+		public Cd (string name, double price):base(name,price)
 		{
 		}
